refactor: converting CSS modules to SCSS modules
