+++ .storybook/main.js
@@ -5,6 +5,7 @@ module.exports = {
     '@storybook/addon-links',
     '@storybook/addon-essentials',
     '@storybook/addon-interactions',
+    '@storybook/preset-scss',
   ],
   framework: '@storybook/react',
   core: {
