+++ components/File/File.module.scss
@@ -1,0 +1,41 @@
+.file {
+  display: grid;
+  place-items: center;
+  width: fit-content;
+  max-width: 150px;
+  margin: 0;
+  cursor: pointer;
+  gap: 10px;
+  $file: &;
+
+  &__icon,
+  &__name {
+    pointer-events: none;
+    user-select: none;
+  }
+
+  &__icon {
+    max-width: 100%;
+    display: block;
+    object-fit: cover;
+  }
+
+  &__name {
+    font-size: 12px;
+    font-family: 'JetBrains Mono', monospace;
+    font-weight: 600;
+    color: black;
+    padding: 6px 10px;
+    text-align: center;
+  }
+
+  &.selected {
+    outline: 2px dashed blue;
+    outline-offset: 8px;
+
+    #{ $file }__icon,
+    #{ $file }__name {
+      background-color: blue;
+    }
+  }
+}
