+++ package.json
@@ -33,6 +33,7 @@
     "@storybook/addons": "^6.5.12",
     "@storybook/builder-webpack5": "^6.5.12",
     "@storybook/manager-webpack5": "^6.5.12",
+    "@storybook/preset-scss": "^1.0.3",
     "@storybook/react": "^6.5.12",
     "@storybook/testing-library": "^0.0.13",
     "@storybook/theming": "^6.5.12",
@@ -48,6 +49,7 @@
     "@typescript-eslint/eslint-plugin": "^5.38.1",
     "babel-loader": "^8.2.5",
     "clsx": "^1.2.1",
+    "css-loader": "^6.7.1",
     "cypress": "^10.9.0",
     "eslint": "8.24.0",
     "eslint-config-next": "12.3.1",
@@ -61,6 +63,9 @@
     "next-translate": "^1.6.0",
     "prettier": "^2.7.1",
     "react-draggable": "^4.4.5",
+    "sass": "^1.55.0",
+    "sass-loader": "^13.0.2",
+    "style-loader": "^3.3.1",
     "ts-jest": "^29.0.2",
     "ts-node": "^10.9.1",
     "typescript": "4.8.4"
