+++ pages/index.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from 'next'
 import Head from 'next/head'
 import Link from 'next/link'
 import Image from 'next/image'
-import styles from '@/styles/Home.module.css'
+import styles from '@/styles/Home.module.scss'
 import useTranslation from 'next-translate/useTranslation'
 import Button from '@/components/Button/Button'
 import { useState } from 'react'
